[CMAT-50] feat : 플래너 생성 다시 추가 및 초기 데이터 타입 수정 (#32)
* [CMAT-50] feat : 플래너 생성 다시 추가 및 초기 데이터 타입 수정

* [CMAT-50] feat:왜래키 관련 오류 수정

* [CMAT-50] fix: 플래너 조회 관련 예외 처리

+++ src/main/java/UMC/career_mate/domain/planner/service/PlannerCommandService.java
@@ -24,21 +24,24 @@ public class PlannerCommandService {
 
     private final PlannerRepository plannerRepository;
 
-    public void savePlanner(Member member){
+    public void initPlanner(Member member){
+
+        if(member == null){
+            throw new GeneralException(CommonErrorCode.NOT_FOUND_BY_MEMBER_ID);
+        }
 
         if(plannerRepository.existsByMember(member)){
             throw new GeneralException(CommonErrorCode.PLANNER_EXISTS);
         }
 
-        List<CreatePlannerDTO> createPlannerDTOList = List.of(
-                CreatePlannerDTO.builder().build(),
-                CreatePlannerDTO.builder().build()
-        );
-
-        plannerRepository.saveAll(PlannerConverter.toPlannerList(member, createPlannerDTOList));
+        plannerRepository.saveAll(PlannerConverter.toInitialPlannerList(member));
     }
 
     public void editPlanner(Member member, CreatePlannerListDTO createPlannerListDTO){
+
+        if(!plannerRepository.existsByMember(member)){
+            throw new GeneralException(CommonErrorCode.PLANNER_NOT_EXISTS);
+        }
 
         List<Planner> planners = plannerRepository.findByMember(member);
         List<CreatePlannerDTO> plannerDTOs = createPlannerListDTO.planners();
@@ -49,6 +52,10 @@ public class PlannerCommandService {
     }
 
     public void deletePlanner(Member member){
+
+        if(!plannerRepository.existsByMember(member)){
+            throw new GeneralException(CommonErrorCode.PLANNER_NOT_EXISTS);
+        }
 
         List<Planner> planners = plannerRepository.findByMember(member);
         plannerRepository.deleteAll(planners);
